requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
enhance==1.2
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.43.1
imageio==2.31.6
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.1
networkx==3.2.1
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
peewee==3.17.0
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
tifffile==2023.9.26
Werkzeug==3.0.1
